Create the EpicsSignalWithRetryRO signal
