+++ src/sophys/common/devices/storage_ring.py
@@ -1,11 +1,12 @@
-from ophyd import Device, Component, EpicsSignalRO
+from sophys.common.utils import EpicsSignalWithRetryRO
+from ophyd import Device, Component
 
 
 class StorageRing(Device):
     """Useful signals from the Storage Ring."""
 
     ring_current = Component(
-        EpicsSignalRO,
+        EpicsSignalWithRetryRO,
         "SI-Glob:AP-CurrInfo:Current-Mon",
         kind="hinted",
         timeout=5,
@@ -16,21 +17,3 @@ class StorageRing(Device):
 
     def __init__(self, *, name, **kwargs):
         super().__init__(prefix="", name=name, **kwargs)
-
-    def describe(self):
-        for _ in range(0, 3):
-            try:
-                get_sts = super().describe()
-                break
-            except Exception:
-                print("Describe Connection Error!! Retrying describe function")
-        return get_sts
-
-    def read(self):
-        for _ in range(0, 3):
-            try:
-                get_sts = super().read()
-                break
-            except Exception:
-                print("Read Connection Error!! Retrying read function")
-        return get_sts
